requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Brotli==1.1.0
cachetools==4.2.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
crispy-bootstrap4==2022.1
cryptography==41.0.4
cssselect2==0.7.0
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-allauth==0.57.0
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-extensions==3.1.5
django-humanize==0.1.2
django-import-export==3.3.1
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-model-utils==4.2.0
django-recaptcha==3.0.0
django-tinymce==3.6.1
django-unicorn==0.44.1
django-weasyprint==2.2.1
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
fontawesomefree==6.4.2
fonttools==4.43.1
gunicorn==21.2.0
html5lib==1.1
humanize==4.8.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
orjson==3.9.10
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.6
pycparser==2.21
pydyf==0.8.0
PyJWT==2.8.0
pyphen==0.14.0
python3-openid==3.2.0
PyYAML==6.0.1
reportlab==4.0.5
requests==2.31.0
requests-oauthlib==1.3.1
shortuuid==1.0.11
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
tinycss2==1.2.1
tzdata==2023.3
urllib3==2.0.6
weasyprint==60.1
webencodings==0.5.1
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
